more modular handling of pipelines
